+++ backend/src/main/java/ssu/groupstudy/domain/comment/service/CommentService.java
@@ -14,8 +14,8 @@ import ssu.groupstudy.domain.comment.repository.CommentRepository;
 import ssu.groupstudy.domain.notice.domain.Notice;
 import ssu.groupstudy.domain.notice.exception.NoticeNotFoundException;
 import ssu.groupstudy.domain.notice.repository.NoticeRepository;
-import ssu.groupstudy.domain.notification.domain.event.NoticeTopicPushEvent;
-import ssu.groupstudy.domain.notification.domain.event.NoticeTopicSubscribeEvent;
+import ssu.groupstudy.domain.notification.domain.event.push.NoticeTopicPushEvent;
+import ssu.groupstudy.domain.notification.domain.event.subscribe.NoticeTopicSubscribeEvent;
 import ssu.groupstudy.domain.user.domain.User;
 import ssu.groupstudy.domain.user.exception.UserNotParticipatedException;
